Major Update - Made the bot MultiGuild

+++ src/commands/keywords.py
@@ -1,6 +1,6 @@
 from bot import bot, discord, pymongo, keywords, time
 from utils.constants import LINK, GUILD_ID, MODLOG_CHANNEL_ID
-from utils.mongodb import kwdb
+from utils.mongodb import kwdb, gpdb
 from utils.roles import is_moderator
 
 
@@ -25,35 +25,36 @@ class AddKeywords(discord.ui.Modal):
         self.add_item(self.autoresponse)
 
     async def callback(self, interaction: discord.Interaction):
-        Logging = bot.get_channel(MODLOG_CHANNEL_ID)
-        timern = int(time.time()) + 1
-        kwdb.add_keyword(
-            self.keyword.value, self.autoresponse.value, interaction.guild.id
-        )
-        keywords[interaction.guild.id] = kwdb.get_keywords(interaction.guild.id)
-        embed = discord.Embed(
-            description="Keyword Created", colour=discord.Colour.green()
-        )
-        embed.set_author(
-            name=str(interaction.user), icon_url=interaction.user.display_avatar.url
-        )
-        embed.add_field(name="Keyword", value=self.keyword.value, inline=False)
-        embed.add_field(
-            name="Auto Response", value=self.autoresponse.value, inline=False
-        )
-        embed.add_field(name="Date", value=f"<t:{timern}:F>", inline=False)
-        embed.add_field(
-            name="ID",
-            value=f"```py\nUser = {interaction.user.id}\nBot = 861445044790886467```",
-            inline=False,
-        )
-        embed.set_footer(text="r/IGCSE Bot#2063")
-        await Logging.send(embed=embed)
-        await interaction.send(
-            f"Created keyword `{self.keyword.value}` for autoresponse `{self.autoresponse.value}`",
-            ephemeral=True,
-            delete_after=2,
-        )
+        mod_log_channel = bot.get_channel(gpdb.get_pref("modlog_channel", interaction.guild.id)) 
+        if mod_log_channel:
+            timern = int(time.time()) + 1
+            kwdb.add_keyword(
+                self.keyword.value, self.autoresponse.value, interaction.guild.id
+            )
+            keywords[interaction.guild.id] = kwdb.get_keywords(interaction.guild.id)
+            embed = discord.Embed(
+                description="Keyword Created", colour=discord.Colour.green()
+            )
+            embed.set_author(
+                name=str(interaction.user), icon_url=interaction.user.display_avatar.url
+            )
+            embed.add_field(name="Keyword", value=self.keyword.value, inline=False)
+            embed.add_field(
+                name="Auto Response", value=self.autoresponse.value, inline=False
+            )
+            embed.add_field(name="Date", value=f"<t:{timern}:F>", inline=False)
+            embed.add_field(
+                name="ID",
+                value=f"```py\nUser = {interaction.user.id}\nBot = 861445044790886467```",
+                inline=False,
+            )
+            embed.set_footer(text="r/IGCSE Bot#2063")
+            await mod_log_channel.send(embed=embed)
+            await interaction.send(
+                f"Created keyword `{self.keyword.value}` for autoresponse `{self.autoresponse.value}`",
+                ephemeral=True,
+                delete_after=2,
+            )
 
 
 class RemoveKeywords(discord.ui.Modal):
@@ -70,28 +71,29 @@ class RemoveKeywords(discord.ui.Modal):
         self.add_item(self.keyword)
 
     async def callback(self, interaction: discord.Interaction):
-        Logging = bot.get_channel(MODLOG_CHANNEL_ID)
-        timern = int(time.time()) + 1
-        kwdb.remove_keyword(self.keyword.value, interaction.guild.id)
-        keywords[interaction.guild.id] = kwdb.get_keywords(interaction.guild.id)
-        embed = discord.Embed(
-            description="Keyword Deleted", colour=discord.Colour.green()
-        )
-        embed.set_author(
-            name=str(interaction.user), icon_url=interaction.user.display_avatar.url
-        )
-        embed.add_field(name="Keyword", value=self.keyword.value, inline=False)
-        embed.add_field(name="Date", value=f"<t:{timern}:F>", inline=False)
-        embed.add_field(
-            name="ID",
-            value=f"```py\nUser = {interaction.user.id}\nBot = 861445044790886467```",
-            inline=False,
-        )
-        embed.set_footer(text="r/IGCSE Bot#2063")
-        await Logging.send(embed=embed)
-        await interaction.send(
-            f"Deleted keyword `{self.keyword.value}`", ephemeral=True, delete_after=2
-        )
+        mod_log_channel = bot.get_channel(gpdb.get_pref("modlog_channel", interaction.guild.id)) 
+        if mod_log_channel:
+            timern = int(time.time()) + 1
+            kwdb.remove_keyword(self.keyword.value, interaction.guild.id)
+            keywords[interaction.guild.id] = kwdb.get_keywords(interaction.guild.id)
+            embed = discord.Embed(
+                description="Keyword Deleted", colour=discord.Colour.green()
+            )
+            embed.set_author(
+                name=str(interaction.user), icon_url=interaction.user.display_avatar.url
+            )
+            embed.add_field(name="Keyword", value=self.keyword.value, inline=False)
+            embed.add_field(name="Date", value=f"<t:{timern}:F>", inline=False)
+            embed.add_field(
+                name="ID",
+                value=f"```py\nUser = {interaction.user.id}\nBot = 861445044790886467```",
+                inline=False,
+            )
+            embed.set_footer(text="r/IGCSE Bot#2063")
+            await mod_log_channel.send(embed=embed)
+            await interaction.send(
+                f"Deleted keyword `{self.keyword.value}`", ephemeral=True, delete_after=2
+            )
 
 
 @bot.slash_command(name="keywords", description="Adds or Deletes a keyword (for mods)")
